Update 15/07/2022 16:59

+++ Netpbm/src/netpbm.hpp
@@ -9,7 +9,7 @@
 #include <bitset>
 
 struct Pixel {
-  int r, g, b;
+  float r, g, b;
 };
 typedef struct Pixel pixel;
